Added a new file up.solvers.results that contains all the new classes implemented to support the PlanGenerationresult class
