+++ .github/workflows/main.yml
@@ -33,7 +33,7 @@ jobs:
         with:
           repository: aiplan4eu/up-tamer
           path: up-tamer
-          ref: "041b45ef32e9190882a01f28bc735540f6d00d1d"
+          ref: "b190389db6f765e7b32bc2b70973d670f295be0f"
 
       - name: Install Tamer unified-planning interface
         run: python3 -m pip install up-tamer/
@@ -43,7 +43,7 @@ jobs:
         with:
           repository: aiplan4eu/up-pyperplan
           path: up-pyperplan
-          ref: "f238ab5c2a6e74721c556d7a82f59ea82e059940"
+          ref: "e10b859baf85c453915b1038f039dea2661feea8"
 
       - name: Install Pyperplan unified-planning interface
         run: python3 -m pip install up-pyperplan/
