+++ unified_planning/solvers/pddl_solver.py
@@ -22,7 +22,7 @@ import subprocess
 import unified_planning as up
 import unified_planning.solvers as solvers
 from unified_planning.shortcuts import *
-from unified_planning.plan import PlanGenerationResult
+from unified_planning.solvers.results import PlanGenerationResult
 from unified_planning.io.pddl_writer import PDDLWriter
 from unified_planning.exceptions import UPException
 from typing import Callable, Optional, List
@@ -66,7 +66,7 @@ class PDDLSolver(solvers.solver.Solver):
                     raise UPException('Error parsing plan generated by ' + self.__class__.__name__)
         return up.plan.SequentialPlan(actions)
 
-    def solve(self, problem: 'up.model.Problem', callback: Optional[Callable[['up.plan.PlanGenerationResult'], None]] = None) -> 'up.plan.PlanGenerationResult':
+    def solve(self, problem: 'up.model.Problem', callback: Optional[Callable[['up.solvers.results.PlanGenerationResult'], None]] = None) -> 'up.solvers.results.PlanGenerationResult':
         w = PDDLWriter(problem, self._needs_requirements)
         plan = None
         log: str = ''
@@ -76,16 +76,15 @@ class PDDLSolver(solvers.solver.Solver):
             plan_filename = os.path.join(tempdir, 'plan.txt')
             w.write_domain(domanin_filename)
             w.write_problem(problem_filename)
-
             cmd = self._get_cmd(domanin_filename, problem_filename, plan_filename)
-            res = subprocess.run(cmd, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-            log = res.stdout.decode()
-            if not os.path.isfile(plan_filename):
-                print(log)
-            else:
+            res = subprocess.run(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            logs: List['up.solvers.results.LogMessage'] = []
+            logs.append(up.solvers.results.LogMessage(up.solvers.results.INFO, res.stdout.decode()))
+            logs.append(up.solvers.results.LogMessage(up.solvers.results.ERROR, res.stderr.decode()))
+            if os.path.isfile(plan_filename):
                 plan = self._plan_from_file(problem, plan_filename)
         status: int = self._result_status(problem, plan)
-        return PlanGenerationResult(status, plan, log_messages=log.split('\n'))
+        return PlanGenerationResult(status, plan, log_messages=logs)
 
     def _result_status(self, problem: 'up.model.Problem', plan: Optional['up.plan.Plan']) -> int: 
         '''Takes a problem and a plan and returns the status that represents this plan.
